moved component to root collection folder as it is used by not only AllCollections but also CollectionListBar
